Change getlocalhost handler
